requirements.txt
easyocr==1.6.2
spacy==3.3.3
segmentation-models-pytorch==0.2.1
attrdict3
pydantic==1.8.2
opencv-python==4.5.5.64
setuptools_git_versioning==1.9.2
setuptools==65.5.1
python-dotenv==0.20.0
wget==3.2
pytesseract==0.3.9
itsdangerous==2.0.1
tensorboard==2.9.1
torch_geometric

#pip3 uninstall -y cryptography
fabric==3.2.1

# numpy>=1.22
# Pillow>=9.1.1
# PyYAML==6.0
# scikit_learn==1.0.2
# 


# 
# gitpython>=3.1.30
# gitdb2==3.0.1
# matplotlib==3.5.2

# fasttext==0.9.2
# networkx==2.8.4
# 

# 
# 

# pandas
